ci: fixes for assorted failing tests
These are grouped together to make merging easier:

* fix: pin lib4sbom to 0.8.2 due to bug (fixes #5031)

This should be temporary until a fix is made upstream.

* fix: test_csv2cve_valid_file test tweak

Number seems to be off at the moment.  May resolve itself but in the interest
of having a working CI I'm just fixing it for now.

* fix: add gcc to augeas test

gcc is being detected in one of the augeas test files, so adding it to the expected other products list.

---------

Signed-off-by: Terri Oda <terri.oda@intel.com>